minor improvement, smaller size for moves in move ordering

+++ inc/c4AI.h
@@ -3,6 +3,7 @@
 
 #include "connect4.h"
 #include "transpositionTable.h"
+#include <cstdint>
 
 class c4AI{
 public:
@@ -29,16 +30,16 @@ private:
     int evaluate_board(connect4& c4, int depth, int alpha, int beta);
     
     // orders the moveOrder array to (hopefuly) be in order of best to worst
-    int order_moves(connect4& c4, int moveOrder[]);
+    int order_moves(connect4& c4, uint8_t moveOrder[]);
 
     //  evaluates the position as it is, from the perspective of the player whose turn it isn't
     int static_evaluate(connect4& c4);
 
     //  a function to be used by an individual thread to search a position
-    //  currently slower, more research needed
+    //  thread search does not work
     void thread_search(connect4 c4, int depth, int threadNum);
 
-    void thread_order_moves(int moveOrder[], int threadNum);
+    void thread_order_moves(uint8_t moveOrder[], int threadNum);
 
 };
